+++ src/blocking_impl.rs
@@ -67,14 +67,10 @@ impl YahooConnector {
 
 /// Send request to yahoo! finance server and transform response to JSON value
 fn send_request(url: &str) -> Result<serde_json::Value, YahooError> {
-    let resp = reqwest::blocking::get(url);
-    if resp.is_err() {
-        return Err(YahooError::ConnectionFailed);
-    }
-    let resp = resp.unwrap();
+    let resp = reqwest::blocking::get(url)?;
     match resp.status() {
-        StatusCode::OK => resp.json().map_err(|_| YahooError::InvalidJson),
-        status => Err(YahooError::FetchFailed(format!("Status Code: {}", status))),
+        StatusCode::OK => Ok(resp.json()?),
+        status => Err(YahooError::FetchFailed(format!("{}", status))),
     }
 }