+++ src/search_result.rs
@@ -40,10 +40,7 @@ pub struct YNewsItem {
 
 impl YSearchResultOpt {
     pub fn from_json(json: serde_json::Value) -> Result<YSearchResultOpt, YahooError> {
-        match serde_json::from_value(json) {
-            Ok(v) => Ok(v),
-            Err(e) => Err(YahooError::DeserializeFailed(e.to_string())),
-        }
+        Ok(serde_json::from_value(json)?)
     }
 }